+++ src/main/scala/com/github/mbuzdalov/oll/BestDynamicIntegerLambda.scala
@@ -29,7 +29,7 @@ object BestDynamicIntegerLambda {
         val tasks = new java.util.ArrayList[Callable[Double]](n)
         for (lambda <- 1 to n) {
           tasks.add(() => ollComputation.findRuntime(parentFitness = x, lambda = lambda,
-            populationSize = lambda, runtimes = runtimes))
+            populationSize = lambda, runtimes = runtimes).toDouble)
         }
         val futures = pool.invokeAll(tasks)
